+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Cards`: Add `--box-shadow` variables on some cards.
+- Fix: `Token`: Update `BoxShadow` tokens values.
 - Feature: `IconBadge`: Add pending status icon.
 - Changes: 
   - `Token`: Rename `--shadow-*` by `--box-shadow-*`.